requirements.txt
python-dotenv
anthropic
requests
praw
huggingface_hub
gspread
google-auth
